requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.36
pydantic==2.9.2
python-dotenv==1.0.1
pyyaml==6.0.2
httpx==0.27.2
anyio==4.4.0
